Rename index-new to index
