+++ api/controllers/StorageController.js
@@ -2,7 +2,7 @@ const PointSDKController = require('./PointSDKController');
 const File = require('../../db/models/file');
 const Chunk = require('../../db/models/chunk');
 const DEFAULT_ENCODING = 'utf-8';
-const {getFile, uploadFile, DOWNLOAD_UPLOAD_STATUS} = require("../../client/storage/index-new.js");
+const {getFile, uploadFile, DOWNLOAD_UPLOAD_STATUS} = require("../../client/storage/index.js");
 
 class StorageController extends PointSDKController {
     constructor(ctx, req) {
