+++ network/web3bridge.js
@@ -6,7 +6,7 @@ const _ = require('lodash');
 const HDWalletProvider = require("@truffle/hdwallet-provider");
 const Web3HttpProvider = require('web3-providers-http');
 const NonceTrackerSubprovider = require('web3-provider-engine/subproviders/nonce-tracker');
-const {getJSON} = require("../client/storage/index-new.js");
+const {getJSON} = require("../client/storage/index.js");
 const ZDNS_ROUTES_KEY = 'zdns/routes';
 const retryableErrors = {
     'ESOCKETTIMEDOUT': 1,
